Refactored away from MWL implementation
